Add readme content and fix 7 row bug
There was a bug where for 7 rows it didn't fit because 60/7 gives a long
decimal answer which doesn't quite fit in 360. So I used Math.floor to remove
this issue. This seems to have solved it without leaving an awkward border.

Also added details of the page in the readme.