Fix black pen bug
The fix in the css for the row alignment issue removed any background colour
from the rows. Which meant when in circile mode, the container background
turned black. I have fixed this by making the row width responsive to the rounded
value of the squares (as rounding down means something they don't occupy
the fully 360).